requirements.txt

# conda install conda-forge::eigen=3.4.0
# export CMAKE_PREFIX_PATH="$CMAKE_PREFIX_PATH:/eigen/path/under/conda"

# # install dependencies
# python -m pip install -r requirements.txt

# # Install NVDiffRast
# python -m pip install --quiet --no-cache-dir git+https://github.com/NVlabs/nvdiffrast.git

# # Kaolin (Optional, needed if running model-free setup)
# python -m pip install --quiet --no-cache-dir kaolin==0.15.0 -f https://nvidia-kaolin.s3.us-east-2.amazonaws.com/torch-2.0.0_cu118.html

# # PyTorch3D
# python -m pip install --quiet --no-index --no-cache-dir pytorch3d -f https://dl.fbaipublicfiles.com/pytorch3d/packaging/wheels/py39_cu118_pyt200/download.html

# # Build extensions
# CMAKE_PREFIX_PATH=$CONDA_PREFIX/lib/python3.9/site-packages/pybind11/share/cmake/pybind11 bash build_all_conda.sh

#fp:pip3 install torch==2.0.0+cu118 torchvision==0.15.1+cu118 torchaudio==2.0.1+cu118 --index-url https://download.pytorch.org/whl/cu118
#mine:pip3 install torch==2.0.1 torchvision==0.15.2 torchaudio==2.0.2+cu118 --index-url https://download.pytorch.org/whl/cu118
#for gdsam2:pip3 install torch==2.3.1 torchvision==0.18.1 torchaudio --index-url https://download.pytorch.org/whl/cu121

# PyTorch and related libraries
# 用下面的下载
#--extra-index-url https://download.pytorch.org/whl/cu118
#torch==2.0.0+cu118
#torchvision==0.15.1+cu118
#torchaudio==2.0.1+cu118

# JupyterLab and widgets
jupyterlab==4.1.5
ipywidgets==8.1.2

# Common Libraries
numpy==1.26.4
scipy==1.12.0
scikit-learn==1.4.1.post1
scikit-image==0.22.0
pyyaml==6.0.1
ruamel.yaml==0.18.6
ninja==1.11.1.1
h5py==3.10.0
numba==0.59.1
pybind11==2.12.0

# CV Libraries
imageio==2.34.0
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
plotly==5.20.0
open3d==0.19.0
#open3d==0.18.0
pyglet==1.5.28
pysdf==0.1.9

# Trimesh
trimesh==4.2.2
xatlas==0.0.9
rtree==1.2.0

# PyRender
pyrender==0.1.45
pyOpenGL>=3.1.0
pyOpenGL_accelerate>=3.1.0

# Meshcat, Webdataset and other utilities
meshcat==0.3.2
webdataset==0.2.86
omegaconf==2.3.0
pypng==0.20220715.0
Panda3D==1.10.14
simplejson==3.19.2
bokeh==3.4.0
roma==1.4.4
seaborn==0.13.2
pin==2.7.0
openpyxl==3.1.2
torchnet==0.0.4
wandb==0.16.5
colorama==0.4.6
GPUtil==1.4.0
imgaug==0.4.0
xlsxwriter==3.2.0
timm==0.9.16
albumentations==1.4.2
xatlas==0.0.9
nodejs==0.1.1
objaverse==0.1.7
g4f==0.2.7.1
#ultralytics==8.0.120
pycocotools==2.0.7
py-spy==0.3.14
pybullet==3.2.6
videoio==0.2.8
kornia==0.7.2
einops==0.7.0
transformations==2024.6.1
joblib==1.3.2
warp-lang==1.0.2

# For sTorch3D
fvcore==0.1.5.post20221221


# SAM and FastSAM related dependencies
xformers==0.0.20
#xformers==0.0.18
torchmetrics==0.10.3
blenderproc==2.6.1
opencv-python
omegaconf
ruamel.yaml
hydra-colorlog
hydra-core
gdown
pandas
imageio
pyrender
#pytorch-lightning==1.8.1
pytorch-lightning==2.3.3
pycocotools
distinctipy
git+https://github.com/facebookresearch/segment-anything.git
ultralytics==8.0.135
timm
gorilla-core==0.2.7.8
#trimesh==4.0.8
gpustat==1.0.0
imgaug
einops

# Download
gdown

###blenderproc安装
#去官网手动安装https://github.com/DLR-RM/BlenderProc
###kaolin==0.15.0 
# 注意torch和CUDA版本
# pip install kaolin==0.15.0 -f https://nvidia-kaolin.s3.us-east-2.amazonaws.com/torch-2.0.1_cu118.html
###pytorch3d==0.7.3
#去官网下载https://anaconda.org/pytorch3d/pytorch3d/files包，然后手动conda install 
###torch==2.0.0+cu118
###torchaudio==2.0.1+cu118
###torchvision==0.15.1+cu118
#pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
